+++ src/components/Menu/Menu.stories.tsx
@@ -4,6 +4,7 @@ import { withDesign } from "storybook-addon-designs";
 
 import Menu, { ListItemsData } from "./Menu";
 import { sectionTypeArray } from "../../helpers/types";
+import { getPlaceholderImage } from "../../utils/utils";
 
 const meta: Meta<typeof Menu> = {
   title: "Components/Navigation/Menu",
@@ -100,7 +101,11 @@ const imageListItems: ListItemsData[] = [
     type: "action",
     id: "item-title-1",
     label: "I'm item 1",
-    media: { type: "image", src: "//placekitten.com/300/300", alt: "" },
+    media: {
+      type: "image",
+      src: getPlaceholderImage("smaller"),
+      alt: "",
+    },
     onClick: () => {
       console.log("Item Title 1 clicked");
     },
@@ -109,7 +114,11 @@ const imageListItems: ListItemsData[] = [
     type: "action",
     id: "item-title-2",
     label: "I'm item 2",
-    media: { type: "image", src: "//placekitten.com/300/300", alt: "" },
+    media: {
+      type: "image",
+      src: getPlaceholderImage("smaller"),
+      alt: "",
+    },
     onClick: () => {
       console.log("Item Title 2 clicked");
     },
@@ -118,7 +127,11 @@ const imageListItems: ListItemsData[] = [
     type: "action",
     id: "item-title-3",
     label: "I'm item 3",
-    media: { type: "image", src: "//placekitten.com/300/300", alt: "" },
+    media: {
+      type: "image",
+      src: getPlaceholderImage("smaller"),
+      alt: "",
+    },
     onClick: () => {
       console.log("Item Title 3 clicked");
     },