+++ src/components/MultiSelect/multiSelectChangelogData.ts
@@ -10,6 +10,20 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Bug Fix",
+    affects: ["Styles"],
+    notes: ["Fixes the z-index value in the small selected items button."],
+  },
+  {
+    date: "2024-04-25",
+    version: "3.1.1",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: ["Updated the placeholder value for the search field."],
+  },
+  {
     date: "2024-03-14",
     version: "3.0.0",
     type: "Update",